Waves Spawning, fish hooking, dice rolling, everything
Waves Spawn when fish is on hook. Both discrete and continuous conditions work. Fish will fuck off the hook if you fail the input accuracy roll, currently set to 60 % accuracy. If you get it correctly, fish is raised out of water, and a function to update UI is called, though it currently does nothing. Has been slightly tested for bugs, hope it works, because the code is actually spaghetti at this point :(
Sorry