Game Manager Prefab
Made Our Nice Game Manager one huge prefab, meaning it saves many of the public variables it needed, so we don't need to plunk them in every time. Is probably gonna not work when we merge, as additional things will be needed, but oh well.

REMEMBER TO PLACE IT IN WORLD SPACE COORDINATES 0, 0, 0!!!