Feedback Update
Added particle effect whenever the player receives feedback for correct input.

Furthermore, if the boat is tilting heavily towards one side in continuous condition, not only will it now slow, but it will change tilting direction. If quick enough on keyboard, should be able to keep boat upright through entire wave.

Finally, made the game force the player to fail the first fish they try to catch in each condition. The first fish would be the one players are most likely to fail, and this guarentees no player is forced to fail 3 times in a row in last quest. Now maximum 2 fails in a row.